+++ lib/marskal/core_ext.rb
@@ -13,6 +13,7 @@ require 'active_support/core_ext/module/deprecation'
 
 require "marskal/core_ext/array"
 require "marskal/core_ext/date"
+require "marskal/core_ext/file"
 # require "marskal/core_ext/my_mod"  #used for testing, take out eventually